requirements.txt
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.147.0
google-generativeai==0.8.3
anthropic==0.39.0
httpx==0.27.2
httpcore==1.0.5
python-docx==1.1.2
schedule==1.2.2
python-dotenv==1.0.1
Pillow==10.4.0
PyPDF2==3.0.1
pymupdf
flask
flask-cors